Add parser for LLVM stackmaps.
This commit implements a parser for LLVM stackmaps which we will later
use to initialise a stopgap interpreter during a guard failure. Unlike
other standard stackmap parsers, this one doesn't create a 1:1
representation of the stackmap as it contains more data than we care
about (for example things like function/const/record counts whose only
purpose is to aid parsing in the stackmap and serve no other purpose
otherwise). Instead, the chosen representation makes it easier to query
the data during a guard failure, e.g. values referred by `ConstIndex`
are folded into `Locations` and records can be retrieved via the return
address of the `__llvm_deoptimize` call, which is calculated during
parsing.