Hook in transition_location properly.
Previously, the JIT could only handle a sole location from which a loop
in the traced program may start (and be compiled etc.).

This change hooks in Laurie's prior `transition_location()` logic,
meaning that we can deal with many loop-starting locations
and (hopefully) in a thread safe manner.

For now the hot threshold is still zero and traces are compiled on the
same thread as the control point. This is because we don't yet have a
way to test the non-determinism that will arise when we change that.
Future PRs should focus on this.