Re-enable parallel compilation.
This change re-enables parallel compilation and adds infrastructure for
tests to serialise compilation via the `YKD_SERIALISE_COMPILATION`
environment variable (which is only available when the `c_testing`
feature is used).

I've yet to devise a test which doesn't use `YKD_SERIALISE_COMPILATION`.
It's not clear if it's possible to test anything meaningful with the
full determinism of threads.

All tests pass, with zero changes to them other than adding the env var.