Remove the inner mutex for the `Compiling` state.
This commit allows the compiling thread to directly alter the
`HotLocation` when compilation is complete, rather than altering a
mutex.

That has the nice property that as soon as compilation is
complete, other threads start executing the compiled code, whereas
before we effectively skipped an iteration of a loop (hence the changes
to the lang_tests).

The less nice property is that we now have to share a raw pointer (to
the HotLocation) amongst threads and make sure that we deal with some
very rare possibilities. In particular, it's possible for a `Location`
to be dropped in one thread while being compiled in another: we have to
introduce a new state to cope with this possibility (which is most
likely to happen if compilation continues after the program has finished
executing).