Pass around `CompiledTrace` by pointer rather than reference.
At the moment we call `CompiledTrace.exec` in the transition function,
so using `&CompiledTrace` is fine. However, we will soon need to pass
such `struct`s outside the transition function, at which point ownership
won't work.

Because of this change, it's important to realise that `CompiledTrace`
objects must be immutable. We weren't mutating them before, so this is
not an onerous change, but it's now explicitly documented.