Remove `ykllvmwrap`'s implicit dependency on ykcapi.
This is a requirement for allowing us to test the `ykrt` state machine
independent of the tracer itself.

As described in #470, any attempt to call `transition_location()` from a
Rust test will currently cause a linkage error (see
https://github.com/ykjit/yk/issues/470#issuecomment-1009829622 for why).

This change solves the linkage problem by avoiding the need for the C++
code in `ykllvmwrap` to call the code in `ykutil`. Instead we call the
`ykutil` code earlier and pass the result in to C++.

Fixes #470.