Minimally rip out the "do something" code in `transition_location`.
`transition_location` both advances the `Location` state machine and
performs actions (e.g. start/stop tracing). Unfortunately, performing
actions makes it difficult to test the state machine in isolation. We
moved one type of action ("execute a trace") out in 0884372d but, as Edd
predicted, that's not enough.

In essence, this commit finishes the job that 0884372 started:
`transition_location` now tells its caller what action to perform
(execute a trace, start/stop tracing, or just nothing).

I have done so as minimally as possible, which means this is somewhat
brutal: I've had to make a number of things `pub` that we would prefer
not to be. Some of those we might be able to de-`pub` at some point
soon. However, to some extent we are a victim of the many `yk` crates,
and to some extent we are a victim of not getting some code into the
"right" crate. I'm not exactly sure what proportion of each category
we'll find we've got, but time will tell!