Remove an incorrect FIXME and a not-very-helpful FIXME.
The first FIXME in this commit is clearly out of date.

The second FIXME isn't out of date, but I don't think this is the right
place for it: we actually talk about what this FIXME is suggesting on
the `YkLocation` struct (and if we want to get into more detail, the
ykdocs repository is probably the right place to do so).