Start work on the stopgap interpreter.
This commit sets up the skeleton of the stopgap interpreter. The interpreter
interacts with the LLVM module via llvm-sys rather than using a more abstracted
(and user-friendly) interface like llvm-ir. The reason is that llvm-ir parses
modules in its entirety, even parts of the module the stopgap interpreter won't
touch. Using llvm-sys we only parse what we actually need to get back to the
control point.

In order to initialise the stopgap interpreter we need to create a mapping from
stackmap values to variables (i.e. instructions) in the AOT module. This is
being setup in jitmodbuilder.cc during guard creation and then passed through
to the stopgap interpreter via the deoptimisation call. We also need to pass
the position in the AOT module from where interpretation needs to start, which
we previously didn't do either.

Note, that the interpreter is still missing a lot of functionality (we can only
interpret branch instructions at the moment).