Add a trace compiler test harness.
This test harness is designed to allow more fine-grained testing of the
trace compiler.

Whereas the existing harness requires whole C programs (interpreter loop
and all), this new suite allows fragments of LLVM IR to be piped through
the trace compiler and the output checked with lang tester.

The user doesn't need to worry about an interpreter loop, or trace
inputs, as those all get faked.