Move from a singleton MT instance to allowing multiple MT instances.
I originally thought of yk as monolithically taking over a process, but
that seems neither necessary nor helpful. For example, there could be
good security reasons to have multiple MT instances within a single
process.

Fortunately, addressing this is very easy: we remove the `MT::global`
method and force people to create MT instnaces via `MT::new`. That's
more-or-less it! It does mean that we need to add a `yk_mt_drop`
function the C API, but that's easily done.