Don't grab the THREAD_MTTHREAD on every control point call.
We can't rely on thread locals being fast (on some platforms they are,
on some they aren't; Rust's thread locals are, currently, not fast
because of the lazy loading) so we want to avoid accessing them in the
control point's fast paths.

This commit moves loading from the `THREAD_MTTHREAD` thread local to
only those portions of code that really need access to it. Mostly this
is fairly simple, but there is one horrible hack in here that will
probably haunt me in my dreams until I can think of a better way of
doing this.

However, getting rid of this thread local should be worth it, as it will
allow us to port back more of the old test suite. At that point,
refactoring the evil hack should be doable with more confidence, because
there'll be a test suite to catch (at least some) errors!