Add a new test to ensure that we don't trace 2 locations in 1 thread.
This highlighted a bug whereby once we'd traced a thread, we forgot to
reset the "this thread is tracing something" setting -- in other words,
each thread could only trace one location in its lifetime.

However, the downside of fixing this is that we can now execute an
"inner" trace while tracing an outer loop. What should we do in such a
situation? As yet, I'm not sure of the best approach...