Add a test for external calls.
I had expected to have to populate the function address map, as
mentioned in a FIXME I had inserted earlier. However, it turns out
that'd only be necessary if we were going to actually load the compiled
trace into the test's memory space (where we'd then get a linker error).

We don't actually do that with trace compiler tests (nor would it make
sense to, since we've faked all of the control point machinery), so the
function address map can in fact remain empty.