Unbreak recursive outlining.
Prior to this change, we'd still emit guards for code that we were
skipping over during the outlining process.

Not only is this unnecessary, it's also incorrect: we may guard the
wrong things. Indeed, this manifests in the (newly reinstated) `fib.c`
test and the (new) `recursion.c` test, where the JIT would stopgap
erroneously!