Make the system honour YKD_PRINT_JITSTATE again.
The JIT state should only be printed if YKD_PRINT_JITSTATE is defined in
the environment.

We used to check for the existence of this variable in ykllvm when
compiling a program. At some point the prints were moved from generated
LLVM IR to Rust code using eprintln!(), but we forgot to check
YKD_PRINT_JITSTATE.

This adds the check back, but (necessarily) the check is done at
runtime, but only if using the "jit_state_debug" feature.

This also pointed out a couple of tests which matched the jit-state
prints, but had not defined YKD_PRINT_JITSTATE.