Check that a "default stand-alone" system doesn't get test/debug features.
During CI check that if we build the system without support for testing
and debugging (i.e. without building any unit tests, without building
the lang_tester suites, and with default features only), that none of
our testing and debugging Cargo features (that impact performance) get
enabled.

At the time of writing, this means that if we run `cargo build -p
ykcapi`, neither the `yk_testing` feature nor the `jit_state_debug`
feature should be used for any crate in the workspace.

We didn't have this right, so the change also contains fixes to make it
the check pass.