Don't allow `MT` to be `Clone`able.
In a sense this is a hangover from "Rust only tracing", when it made
sense to make this feel very Rust-y, with interior mutation. That no
longer feels like a great design decision as a lot of code that
interacts with this won't be in Rust, so it means that we pay an
internal indirection penalty (for `MTInner`) that languages such as C
don't need to pay.

Forcing users to put an `MT` instance in an `Arc` if they need Rust (and
multi-threaded Rust at that!) feels like it's worth the API consistency.