Fix a bug where we may trace execution of a trace.
In doing so, make the interp_loop tests more realistic (don't have
locations for bytecode positions where a loop can't possibly start).

Note that this gives us correctness, but is not really what we want in
the long run.

The situation we are talking about here is essentially where there are
nested loops in the interpreter. The inner loop will be JIT compiled
first, then later the outer loop will eventually be compiled too. It
would be best if the outer loop called the existing trace for the inner
loop, rather than duplicating JITted code for the inner loop.

We will attempt to implement the correct behaviour later.

Co-authored-by: Laurence Tratt <laurie@tratt.net>