Add an assertion and fix a failure of the assertion.
When investigating another issue, I noticed that we sometimes resume
from inlining a call at a block different from that in which the call
happened.

For example, we may have AOT IR like:
```
bb3:
  call @f()
  br bb4

bb4:
  %8 = add i32...
```

And we may see a trace like:
<bb3, fbbs, bb4>

instead of the expected:
<bb3, fbbs, bb3, bb4>

(here `fbbs` represents one or more block from `@f()`).

The trace compiler expects to see `bb4` when returning from `@f()`, as a
`call` instruction doesn't terminate a high-level LLVM block.

For this example, the problem is that when lowering to machine IR,
SelectionDAG-based lowering doesn't emit anything for `br bb4` because
the blocks are consecutive and control flow can simply fall through.
This, however, messes with the mapping process of our JIT, which
requires the machine IR to have the same block structure as the
high-level LLVM IR.

For now we first add an assertion on the `yk` side to catch "unexpected
resumes". This causes existing tests to fail as-is, so to fix this we
next add a ykllvm flag `--yk-no-fallthrough`, which ensures that the
branch is always emitted, even if fallthrough is possible.

Side-note: In this case, we *might* have been able to recover from
unexpectedly resuming from `@f()` at `bb4` by checking that the
terminating instruction of the "expected block" (`bb3`) was an
unconditional jump to `bb4`. However, it's not clear if this strategy is
correct in general. In light of that we opt to do something obviously
correct, rather than maybe faster but potentially wrong.