Start work on the stopgap interpreter.
The stopgap interpreter is called during a guard failure and interprets
AOT IR until it reaches a point where the compiled interpreter can take
over (i.e. when reaching the control point again) or until it interprets
the return of the main interpreter loop.

The interpreter interacts with the LLVM module via llvm-sys rather than
using a more abstracted (and user-friendly) interface like llvm-ir. The
reason is that llvm-ir parses modules in its entirety, including parts
of the module the stopgap interpreter won't touch. Using llvm-sys we
only parse what we actually need to get back to the control point.

In order to initialise the stopgap interpreter we need to create a
mapping from stackmap values to variables (i.e. instructions) in the AOT
module. This is being setup in jitmodbuilder.cc during guard creation
and then passed through to the stopgap interpreter via the
deoptimisation call. We also need to pass the position in the AOT module
from where interpretation needs to start, which we previously didn't do
either.

In order to exit properly after interpreting a return from the main
interpreter loop, the stopgap interpreter returns an integer to tell the
compiled interpreter whether it has reached the control point or
interpreted a return so that it can act accordingly (see ykllvm
changes).

Note, that this is still missing some functionality, e.g. stack frames,
global variables, and takes some liberties to get the tests working
(e.g. skipping some function calls). These will be fixed separately.