Make the trace compiler actually optimise.
Due to incorrect usage of the pass manager, our trace compiler has not
been optimising traces!

This commit does 3 things.

First it makes the trace compiler optimise at -O2, as was intented from
the outset.

After making this change it became immediately obvious that
optimisations were occurring: the whole JIT function was removed because
we had it marked for internal linkage. So the second thing this commit
does is mark the JIT function as `ExternalLinkage`.

And finally, now that we optimise, items in the global mappings could be
optimised out. For such cases we shouldn't register global mappings for
them, or LLVM assertions will fail.