Add "runtime filtering" to the trace compiler test suite.
If an `.rtd` file is present alongside a test file, then the `Run-time`
stage of the lang test is replaced by a `Run-time-filtered` stage, whose
text input is acquired by executing the `.rtd` file.

The `.rtd` file will typically be a shell script (but it could be
anything executable) and is passed as arguments:
- The path to the `run_trace_compiler_test` binary.
- The path to the `.ll` file being tested.

The filter can then do anything it wants to generate the output text to
be matched in the main test file's `Run-time-filtered` specification.

This allows us to do more flexible things with our tests. For example
I've included a `guards_collapse.ll` test which checks that subsumed
guards are collapsed (optimised out). The test works by munging the
output into a form that allows us to (concisely) check for the *absence*
of the optimised out guard failure: something which cannot easily be
achieved with lang_tester alone.

Put differently, by putting the text into a form that matches only what
we want and without the need to used `...` match lines (that would
potentially consume the lines we want to check *don't* exist), we are
able to check for the absence of certain lines in test output.

This only implements runtime filtering for the trace compiler suite for
now, but we can easily ad support to the C test suite later on-demand.