Fix stopgapping out of nested functions.
Until now our stopgap interpreter was only able to recover from a guard
failure if it happened at the top level of the compiled interpreter.
This commit adds the neccessary extra information to the guards so that
the interpreter can recover even when the guard failure happens deep
inside a nested (and in the JIT inlined) function.