Teach the stop-gap interpreter how to call AOT functions.
When encountering a call instruction, we don't interpret the callee, but
instead call the AOT version of the function. This should be faster.

We do this via libffi bindings to Rust, as we need the ability to call
addresses whose interfaces not known until runtime.

In doing so we add some more wrappers to the LLVM bridge.