Remove the TracerState enum.
This is currently just a bool in disguise, so removing it, and using an
actual bool simplifies the code noticeably. I also think that it's
plausible that some backends might one day have a more sophisticated
notion of their internal state that can't be captured by a generic enum,
so it's best to let backends choose how they record and expose that
state.