Deoptimise minimally needed live values.
The previous algorithm to collect live values for deoptimisation inside
a trace overapproximated so much that large traces could take minutes to
build. This was due the linear nature of the trace which meant no longer
used values would stay live (from the perspective of the algorithm):
the bigger the trace the more values would end up in the later
guard failures.

Instead of the overapproximation we now lookup live values inside the
AOT module, using the recently added stackmap calls. At each guard
failure, we lookup the corresponding stackmap call inside AOTMod,
extract the live values, and translate them to their corresponding
JITMod values. These are the exact live values that we need to
deoptimise.

This change shook out a bug with the liveness analysis in ykllvm (which
is used to calculate stackmap call operands). Since we now deoptimising
exactly what the AOT stackmap calls tell us, it turned out that we also
need to include PHINode operands in the liveness analysis which we
previously skipped. This has been fixed in
https://github.com/ykjit/ykllvm/pull/41/commits/4c1baeb83137ac2ec31f61742f2291d752d7b62d.