Don't use the LLVM symbolizer to look up symbols.
Part of the mapping phase of JIT requires the ability to map a block's
virtual address to the symbol from which it came.

Before we were using the `llvm::LLVMSymbolizer` API for this, but
recently when debugging an issue, we found that this API favours debug
info over the symbol table. This means that each time we map a block, we
look for DWARF debug info, and that seems a little wasteful and
questionably accurate.

This change therefore switches to using `dladdr(3)` instead.

Using the test suite as a kind of benchmark, running the tests with a
single thread like this:
```
multitime -n 10 env RUST_TEST_THREADS=1 cargo test --release
```

Before this commit:
```
1: env RUST_TEST_THREADS=1 cargo test --release
            Mean        Std.Dev.    Min         Median      Max
real        38.131      0.636       36.739      38.260      39.202
user        260.719     0.979       259.012     260.509     262.621
sys         561.254     1.084       559.982     561.146     563.089
```

After this commit:
```
1: env RUST_TEST_THREADS=1 cargo test --release
            Mean        Std.Dev.    Min         Median      Max
real        36.826      0.775       35.561      37.182      38.165
user        252.629     0.911       250.878     252.606     254.124
sys         541.665     2.248       538.241     542.094     544.800
```

(Note that 3 new tests were added, so this isn't quite apples to apples)

Although the speedup isn't dramatic, the code is simplified quite a bit,
so I think this is worthwhile.

I also think that this could be sped up further with caching, but that's
one for another day.