Ensure that hwtracer can decode PT traces in the face of chdir().
Without this change, if the process to be traced was invoked with a
relative path and changes directory before decoding a trace, then the PT
backend would try to create a libipt image using a (stale) relative path
for the main binary's object.

As discussed with @ltratt, we fix this is a hacky fashion for now, using
current_exe() to get an absolute path to the executable file. We
shouldn't rely on this in the long term.