Remove higher optimisation levels.
For now we can only compile with -O0 since higher optimisation levels
introduce machine code we currently don't know how to deal with, e.g.
temporaries which break stackmap reconstruction. This isn't a huge problem
as in the future we will keep two versions of the interpreter around and
only swap to -O0 when tracing and run on higher optimisation levels
otherwise.