Add a `yk-config` script.
This a is little shell script loosely based around the various other
`*-config` tools that you see in the unix ecosystem.

The goal is to eliminate the burden of providing the right compiler
flags when making an interpreter use the Yk JIT. The interpreter's build
system should invoke our script to get the flags.

It's still a bit rough around the edges. For example you have to tell
the script if we are using release or debug build of yk so that it can
find libykcapi.so. I imagine that in the long-term, yk would be
installed properly (e.g. in /usr/local) at which point we won't be
linking to shared objects in a Rust target directory and this hack could
probably go away. Regardless, this will do for now.