Make JITted trace code debuggable.
This change adds an environment variable `YKD_TRACE_DEBUGINFO`, which
when set to "1" causes the JIT to add debug locations to the JITMod.

This allows debuggers conforming to the gdb JIT interface to display
higher-level representations of the JITted machine code as it is being
debugged.

The "source code" displayed is a temporary file generated on the fly. It
contains jit-post-opt IR annotated with AOT source code locations.

The temporary file is managed by a `tempfile::NamedTempFile` which is
stored in `CompiledTrace`. This means that when we finally get a
mechanism to evict old traces, the temp file will be deleted as the
compiled trace is dropped.

In order to show good quality output, I've also had to give the blocks
in the JITMod meaningful names. This caused some test flux.

Some of the code here is UNIX specific, but I've guarded the sections
that would have to be adapted for other OSs.

C++17 is required for `std::filesystem`.