Reconstruct stack during guard failures.
Until now we've interpreted AOT IR after a guard failure to get back to
the control point. While this meant we needed write a complete LLVM IR
interpeter - already no simple feat - the bigger problem was that it was
unlikely this interpreter was ever going to be fast.

Unfortunately, the only alternative sounded scary: reconstruct the stack
in a way as if we've never executed a trace so we can immediatley jump
back into compiled code after a guard failure without the need for
interpretation. Luckily for us LLVM's stackmaps make this solution
slightly less scary, and with @vext01's recent stackmap changes in LLVM
upstream we were finally ready to tackle this.

It took a while, but the change is finally here. It works as follows:
when a guard fails we call `llvm.experimental.deoptimize` as before,
which in return calls `yk_stopgap`. This then parses the mapping from
JIT to AOT variables, and initialises the stopgap interpreter (as
before) with the correct values. However, instead of interpreting AOT,
we now call `reconstruct_frames`, where we assemble a new stack in
memory, which will later overwrite the current stack. The function then
returns the address of the new stack, all the way back to the
`control_point`'s caller. If the control point returns anything but a
nullpointer we call `__ykrt_reconstruct_frames` which handles the
replacement of the current stack. Once this is done it automatically
jumps to the address in the AOT code which resulted in the guard failure
and continues running the program as normal.