Separate trace collection from trace decoding.
We are about to add our own trace decoder (as an alternative to libipt)
to hwtracer. In order to do this we need the ability to have separate
notions of trace collection and trace decoding.

In other words, instead of having a `PerfPT backend`, we now have a
`PerfTraceCollector` and a `LibIPTTraceDecoder`.

(There was also something called a `Dummy` backend, which I saw no
utility to, and have thus removed)

This is a large change, that will be hard to review I'm afraid.

In doing this work I've realised how crusty and unidiomatic some of the
hwtracer code is. Regardless I've tried hard to resist the urge to
uncrustify unrelated parts of code. I will fix some of this stuff in
follow up PRs.