Run tests for both PT decoders.
We run the C suite twice, once for each PT decoder.

(It doesn't make sense to do so in the other suites as they either
explicitly test one specific decoder, or don't use PT to get a trace)

In order to select which decoder to use we introduce the
YKD_FORCE_TRACE_DECODER env var. This is strictly a debugging/testing
facility. Eventually we will add an API that allows the interpreter
authors to select a decoder if they wish (but not now).