Work around common FUP interruptions.
This works around the (fairly common) case where normal execution is
temporarily interrupted by code external to the interpreter itself while
we are collecting a trace. Such interruptions cause a FUP packet in the
packet stream. We simply ignore the interruption.

Later we need to rethink how we deal with FUPs as the interpreter could
interrupt itself, and ignoring that will be harder. See:
https://github.com/ykjit/yk/issues/593