Handle FUP/TIP.* packets that pair with MODE.* packets.
The Intel manual explains that if MODE.* packets precede a TIP.* or a
FUP packet, then the FUP packet is consumed.

These cases communicate the execution mode of the CPU (WRT 16/32/64-bit
mode, and transaction state) and are emitted when the mode changes, but
also in PSB+ sequences (a decoder may have missed a prior mode change
due to being desynchronised)

This differs from the case (that we've already partially handled), where
a FUP packet indicates an interruption in the regular control flow of
the instruction stream.

This change ensures that we can handle MODE changes. Fixes one class of
non-deterministic crashes.

While here, properly parse the fields of MODE.* packets and check that
we only ever see 64-bit mode.