Fix a rare assertion failure.
The FUP workaround code checks for a sequence of three PT packets that
indicate that the interruption was an "easy workaround case".

In the case that the third packet happens to be a `TIPPGD`, the
decoder's `self.pge` was not correctly updated. This meant that when the
decoder sees the corresponding `TIPPGE` and it asserts `!self.pge`, the
assertion fails.

This change reshuffles the code a bit so that the update to `self.pge`
always happens correctly.