Simplify the early stages of the trace compiler.
Before, the trace compiler had a notion of "have we encountered the
first call to the control point yet?". It tracked this with some state
called `NewControlPointCall`, which became a non-null pointer when it
saw the control point. It used this to decide when to stop ignoring
instructions before the first call to the control point, and when to
start actually copying instructions into the JITMod.

Suppose the first block of the trace (which should always contain the
call to the control point, by the way) is:

 blah1
 blah2
 control_point(...)
 blah3
 blah4

The trace compiler would iterate over `blah1` and `blah2` in its main
loop with `NewControlPointCall` set to NULL, not copying those
instructions into `JITMod`. After seeing the call to the control point,
`NewControlPointCall` is set non-null and `blah3` and `blah4` will be
copied in subsequent loop iterations.

The problem with this approach is that there are many places where we
have to check if we have seen the first call to the control point before
doing something. This only adds complexity to the already brutally
complex trace compiler.

In addition, becuase the compiler would start from the beginning of the
first block of the trace (not after the control point), it would think
that it was seeing the call to the control point, which is impossible,
as tracing couldn't have started. Some sloppy `ExpectUnmappable`
handling was letting this through and "working by chance" (tm).

This change "pre-seeks" the trace compiler loop to start from directly
after the control point, meaning that the trace compiler's main loop can
always assume that the control point has already been seen, and that it
should start copying instructions straight away. We achieve this using
the existing `ResumeAfter` facility.

Further, we tighten up the `ExpectUnmappable` handling.

The upshot of this is that the code is simplified quite a bit.

No functional change.