Make the mapper eager.
Prior to this change, the mapper was a pseudo-iterator with a `next()`
but using a non-standard interface requiring the consumer to pass in the
last value the iterator yielded. This was required in order to collapse
consecutive unmappable blocks, and to omit the initial unmappable prefix
(both assumptions the trace compiler makes).

This interface is not only odd, but it is hindering my work on allowing
foreign code to call back to mappable code.

This change makes the mapper eagerly collect the mapped trace, instead
of working like an iterator. By doing so, we kill the weird interface
and simplify the code quite a bit.

At first I was hesitant to make this change: I'm biased towards
iterators, as they can often be made to use less memory than collecting.
But upon inspection, both consumers of this interface (yktrace and one
of the test suites) simply collect the trace anyway, so there's really
no need for the mapper to be an iterator.