Refactor the trace compiler.
The trace compiler contained a lot of state that was all basically a
different view on the same callstack information. The sheer amount of
scattered state made the code hard to understand and maintain.

This change encapsulates all of that state into one easy to understand
data structure: `CallStack`.

Notes:
 - Before this change, the initial frame was not modelled. After this
   change, it is. This makes the code more consistent and easier to
   comprehend.

 - Before, foreign frames were not modelled. Now they are. This will
   play into the "allowing foreign code to callback" work that will
   follow.

Despite all this: no functional change!