Allow foreign code to call back to mappable code.
Before we disallowed foreign code to "call back" to mappable interpreter
code. For example, this would be disallowed:

```
int cmp(const void *a, const void *b) { ... }

int main(void) {
    ...
    qsort(..., cmp);
    ...
}
```

In the above `qsort` is foreign code not compiled by ykllvm that we
don't have IR for. Until now we'd assume that foreign code doesn't call
back to mappable (not foreign) code, and that it always returns directly
back to mappable code.

This change allows us to differentiate returning-to-mappable from
calling-back-into-mappable, handling the two scenarios correctly.

This works by tracking the stack effects of unmappable code during
instruction disassembly and either popping-from or pushing-to the trace
compiler's abstract stack based on this value.

A signed integer is use to track the relative stack height: +N means the
stack is deeper by N frames, whereas -N means the stack is shallower by
N frames.

In order for this to work, I had to give the ykpt decoder iterator a
one-element lookahead. This is because at the time the unknown block
element is created, we don't yet know the stack effect of the unmappable
code: we will only know that once we have disassembled up until we get
back to mappable code. By caching the next element, we can patch it up
before the iterator gives it out.

This only works for the ykpt decoder for now. My next unit of work will
be to try to get this working for libipt, but it requires quite a bit of
refactoring that would clutter up this PR.