Fix several classes of non-deterministic failure.
Periodically the PT hardware emits a PSB+ packet sequence. Each of these
consists of:

 - A PSB packet.
 - other packets.
 - A PSBEnd packet.

The Intel manual (Section 33.3.7) explains that the "other packets" are special
and don't conform to the normal rules:

> When a PSB packet is generated, it is followed by a PSBEND packet (Section
> 33.4.2.18). One or more packets may be generated in between those two
> packets, and these inform the decoder of the current state of the
> processor. These packets, known collectively as PSB+, should be interpreted
> as “status only”, since they do not imply any change of state at the time
> of the PSB, nor are they associated directly with any instruction or event.
> Thus, the normal binding and ordering rules that apply to these packets
> outside of PSB+ can be ignored when these packets are between a PSB and
> PSBEND. They inform the decoder of the state of the processor at the time
> of the PSB.

Our mistake was to allow target ip payloads in these "other packets" to update
our decoder state, namely `self.cur_loc`. In some cases this would make it
appear as though certain blocks were being re-executed in part. In turn this
confused later parts of the JIT pipeline.