Update bottom-most frame during deoptimisation.
The bottom-most frame during deoptimisation is the frame from which the
trace was run. This frame already exists at deopt time and thus doesn't
need to be recreated. We falsely assumed this also meant we don't need
update its values. However, the JIT trace may temporarily spill values to
its own stack, then deoptimise, and then reload those values. In this case
we need to copy the trace stack values over to the AOT stack.