Reduce the LLVM builds in CI from 3 to 2.
I believe that we were building 3 versions of LLVM in CI: one in
`.buildbot.sh` and once each for `cargo test` and `cargo test
--release`.

This commit gets this down to 2 LLVM builds: an LLVM-with-assertions
which we use with the relevant `YKB` env var; and an
LLVM-without-assertions for `cargo bench`. The latter is also intended
to test yk's ability to build ykllvm without having the env var set, so
kills two birds with one stone.